Observable->StreamResult and more
