+++ clients/ts/signalr/spec/Common.ts
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-import { ITransport, HttpTransportType } from "../src/ITransport";
+import { HttpTransportType, ITransport } from "../src/ITransport";
 
 export function eachTransport(action: (transport: HttpTransportType) => void) {
     const transportTypes = [
