+++ clients/ts/signalr/spec/HttpConnection.spec.ts
@@ -1,11 +1,10 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-import { DataReceived, TransportClosed } from "../src/Common";
 import { HttpConnection } from "../src/HttpConnection";
 import { IHttpConnectionOptions } from "../src/HttpConnection";
 import { HttpResponse } from "../src/index";
-import { ITransport, TransferFormat, HttpTransportType } from "../src/ITransport";
+import { HttpTransportType, ITransport, TransferFormat } from "../src/ITransport";
 import { eachEndpointUrl, eachTransport } from "./Common";
 import { TestHttpClient } from "./TestHttpClient";
