+++ clients/ts/signalr/src/Loggers.ts
@@ -11,30 +11,3 @@ export class NullLogger implements ILogger {
     public log(logLevel: LogLevel, message: string): void {
     }
 }
-
-export class ConsoleLogger implements ILogger {
-    private readonly minimumLogLevel: LogLevel;
-
-    constructor(minimumLogLevel: LogLevel) {
-        this.minimumLogLevel = minimumLogLevel;
-    }
-
-    public log(logLevel: LogLevel, message: string): void {
-        if (logLevel >= this.minimumLogLevel) {
-            switch (logLevel) {
-                case LogLevel.Error:
-                    console.error(`${LogLevel[logLevel]}: ${message}`);
-                    break;
-                case LogLevel.Warning:
-                    console.warn(`${LogLevel[logLevel]}: ${message}`);
-                    break;
-                case LogLevel.Information:
-                    console.info(`${LogLevel[logLevel]}: ${message}`);
-                    break;
-                default:
-                    console.log(`${LogLevel[logLevel]}: ${message}`);
-                    break;
-            }
-        }
-    }
-}
